trying to put the dialog out of the main file
